refactor(repo): move public methods
