+++ repo/fsrepo/fsrepo.go
@@ -6,10 +6,11 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/jbenet/go-ipfs/repo"
+	repo "github.com/jbenet/go-ipfs/repo"
+	common "github.com/jbenet/go-ipfs/repo/common"
 	config "github.com/jbenet/go-ipfs/repo/config"
 	util "github.com/jbenet/go-ipfs/util"
-	"github.com/jbenet/go-ipfs/util/debugerror"
+	debugerror "github.com/jbenet/go-ipfs/util/debugerror"
 )
 
 type FSRepo struct {
@@ -105,6 +106,42 @@ func (r *FSRepo) SetConfig(conf *config.Config) error {
 	}
 	*r.config = *conf // copy so caller cannot modify the private config
 	return nil
+}
+
+// GetConfigKey retrieves only the value of a particular key
+func (r *FSRepo) GetConfigKey(key string) (interface{}, error) {
+	filename, err := config.Filename(r.path)
+	if err != nil {
+		return nil, err
+	}
+	var cfg map[string]interface{}
+	if err := readConfigFile(filename, &cfg); err != nil {
+		return nil, err
+	}
+	return common.MapGetKV(cfg, key)
+}
+
+// SetConfigKey writes the value of a particular key
+func (r *FSRepo) SetConfigKey(key string, value interface{}) error {
+	filename, err := config.Filename(r.path)
+	if err != nil {
+		return err
+	}
+	var mapconf map[string]interface{}
+	if err := readConfigFile(filename, &mapconf); err != nil {
+		return err
+	}
+	if err := common.MapSetKV(mapconf, key, value); err != nil {
+		return err
+	}
+	if err := writeConfigFile(filename, mapconf); err != nil {
+		return err
+	}
+	conf, err := convertMapToConfig(mapconf)
+	if err != nil {
+		return err
+	}
+	return r.SetConfig(conf)
 }
 
 func (r *FSRepo) Close() error {
